requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==24.1.0
azure-core==1.30.2
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-schema==4.16.1
botframework-connector==4.16.1
botframework-streaming==4.16.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
frozenlist==1.4.1
idna==3.7
isodate==0.6.1
jsonpickle==1.4.2
msal==1.30.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.19
yarl==1.9.4